+++ service-implementation/src/main/java/com/liaison/mailbox/service/glass/util/TransactionVisibilityClient.java
@@ -8,7 +8,7 @@
  * with Liaison Technologies.
  */
 
-package com.liaison.mailbox.service.util;
+package com.liaison.mailbox.service.glass.util;
 
 import java.util.Date;
 
@@ -24,6 +24,7 @@ import com.liaison.commons.util.settings.DecryptableConfiguration;
 import com.liaison.commons.util.settings.LiaisonConfigurationFactory;
 import com.liaison.mailbox.MailBoxConstants;
 import com.liaison.mailbox.enums.ExecutionState;
+import com.liaison.mailbox.service.util.MailBoxUtil;
 
 /**
  * Java wrapper client for logging messages in LENS.