+++ service-implementation/src/main/java/com/liaison/mailbox/service/util/KMSUtil.java
@@ -41,6 +41,7 @@ public class KMSUtil {
 
 	private static final Logger LOGGER = LogManager.getLogger(KMSUtil.class);
 	public static final String PROPERTY_KEY_MANAGEMENT_BASE_URL = "client.key-management.baseUrl";
+	public static final String UTF_ENCODE = "UTF-8";
 
 	/**
 	 * Util method get stored secret from KMS
@@ -59,7 +60,7 @@ public class KMSUtil {
 		// setting the request headers in the request to key manager from gem
 		Map<String, String> headerMap = GEMHelper.getRequestHeaders(gemManifestFromGEM, "application/json");
 
-		String url = MailBoxUtil.getEnvironmentProperties().getString("kms-base-url") + "secret/" + URLEncoder.encode(guid, "UTF-8");
+		String url = MailBoxUtil.getEnvironmentProperties().getString("kms-base-url") + "secret/" + URLEncoder.encode(guid, UTF_ENCODE);
 		String base64EncodedPassword = HTTPClientUtil.getHTTPResponseInString(LOGGER, url, headerMap);
 
 		if (base64EncodedPassword == null || base64EncodedPassword.isEmpty()) {
@@ -90,7 +91,7 @@ public class KMSUtil {
 		url = url + "fetch/group/keypair/current/";
 
 		// To be fetched from DataBase
-		url = url + keypairPguid;
+		url = url + URLEncoder.encode(keypairPguid, UTF_ENCODE);
 		// get gem manifest response from GEM
 		GEMManifestResponse gemManifestFromGEM = GEMHelper.getACLManifest();
 
@@ -129,7 +130,7 @@ public class KMSUtil {
 		url = url + "fetch/truststore/current/";
 
 		// To be fetched from DataBase
-		url = url + trustStoreId;
+		url = url + URLEncoder.encode(trustStoreId, UTF_ENCODE);
 
 		// get gem manifest response from GEM
 		GEMManifestResponse gemManifestFromGEM = GEMHelper.getACLManifest();